Begin splitting dtach into create/attach commands
The dtach command has two modes which are almost completely disjoint:
create and attach.  Splitting these gives us greater flexibility and the
ability to add specific options to one or the other.

For example, `dtach -c socket -z' is equivalent to:
 dtmaster socket -w && dtattach socket -z

As of this commit, the commands only print their usage.  This will
change, of course!