fix return value for nofork
Seems that closing the last pty slave generates an EIO on the master.
For real serial TTYs, this would probably mean there was an I/O error.
In our case, it's pretty much the condition for normal termination, so
return 0 when we get EIO.

Eventually want to rewrite this to use SIGCHLD as the termination notice
rather than EIO...